Move test function to a testing file
